[skip] Updated sqlite3 conservatively (4m)
Admittedly, this does not LOOK like a conservative update. I needed to bump sqlite3 to 1.3.13 to fix a bug compiling it on High Sierra; but the newer version of sqlite changed its "RecordNotUnique" exception; and I needed to do a patchlevel upgrade of Rails to get the corresponding fix to ActiveRecord.