[fix] Removed ThreadsafeAttributes (4h)
ThreadsafeAttributes was a good idea. However, its implementation was not a good fit for how we need state to work in Remotable: When a threadsafe attribute is first set, it sets the value of the variable on the current thread and then _also_ on the main thread, if the variable is not yet defined on the main thread. There were two different situations where this was a bad idea for remotable:

First, when doing class initialization, defaults are set for several attributes in the `ActiveRecordExtender`, with the intent that the consuming class would be able to set their own values later. However, if the class initialization is not run on the main thread, setting the defaults in `ActiveRecordExtender` would set the default value on both the current and the main threads, but the override that followed would only be set on the current thread. This led to only one thread having the correct, intended value for that attribute for the model.

The second situation was when setting a value temporarily for the span of a block, and then setting the value back afterwards. Again, assuming the wrapping method was called on a non-main thread, it would set that override value for the current thread _and_ the main thread, if the main thread was not yet defined, but then the call to set it back to its original value after the block had executed would _only_ reset the value on the non-main thread, since we just ensured the value was defined on the main thread before running the block.

The solution, then, was twofold. For the first situation, since the intent is to set values tied to the actual model, which are meant to be global, we simply reverted to using class variables, which are set across threads. For the second situation, we manually wrapped thread variables that only applied to the current thread and didn't set anything on the main thread. By limiting scope in this way, we prevented polluting the main thread with values intended to be temporary.