[fix] Removed deprecated calls to `URI.encode` (5m)
`URI.encode` and `CGI.encode` are not entirely interchangeable. It would seem that the former doesn't escape slashes, etc. while the latter is more aggressive in escaping everything, but replaces spaces with `+` instead of `%20`. I opted to go a different route with `ERB::Util.url_encode`, which aggressively escapes everything BUT also uses `%20` when escaping spaces, which is most consistent with current behavior in remotable. I don't love having to have updated the test output in this case, but it _should_ still work with any real server, even if the actual URL produced is different.